+++ invokeai/frontend/web/src/features/nodes/util/graph/addNSFWCheckerToGraph.ts
@@ -2,6 +2,7 @@ import type { RootState } from 'app/store/store';
 import type { ImageNSFWBlurInvocation, LatentsToImageInvocation, NonNullableGraph } from 'services/api/types';
 
 import { LATENTS_TO_IMAGE, NSFW_CHECKER } from './constants';
+import { getBoardField, getIsIntermediate } from './graphBuilderUtils';
 
 export const addNSFWCheckerToGraph = (
   state: RootState,
@@ -21,7 +22,8 @@ export const addNSFWCheckerToGraph = (
   const nsfwCheckerNode: ImageNSFWBlurInvocation = {
     id: NSFW_CHECKER,
     type: 'img_nsfw',
-    is_intermediate: true,
+    is_intermediate: getIsIntermediate(state),
+    board: getBoardField(state),
   };
 
   graph.nodes[NSFW_CHECKER] = nsfwCheckerNode as ImageNSFWBlurInvocation;
