+++ invokeai/frontend/web/src/features/nodes/util/graph/buildAdHocUpscaleGraph.ts
@@ -1,51 +1,33 @@
-import type { BoardId } from 'features/gallery/store/types';
-import type { ParamESRGANModelName } from 'features/parameters/store/postprocessingSlice';
-import type { ESRGANInvocation, Graph, LinearUIOutputInvocation, NonNullableGraph } from 'services/api/types';
+import type { RootState } from 'app/store/store';
+import { getBoardField, getIsIntermediate } from 'features/nodes/util/graph/graphBuilderUtils';
+import type { ESRGANInvocation, Graph, NonNullableGraph } from 'services/api/types';
 
-import { ESRGAN, LINEAR_UI_OUTPUT } from './constants';
+import { ESRGAN } from './constants';
 import { addCoreMetadataNode, upsertMetadata } from './metadata';
 
 type Arg = {
   image_name: string;
-  esrganModelName: ParamESRGANModelName;
-  autoAddBoardId: BoardId;
+  state: RootState;
 };
 
-export const buildAdHocUpscaleGraph = ({ image_name, esrganModelName, autoAddBoardId }: Arg): Graph => {
+export const buildAdHocUpscaleGraph = ({ image_name, state }: Arg): Graph => {
+  const { esrganModelName } = state.postprocessing;
+
   const realesrganNode: ESRGANInvocation = {
     id: ESRGAN,
     type: 'esrgan',
     image: { image_name },
     model_name: esrganModelName,
-    is_intermediate: true,
-  };
-
-  const linearUIOutputNode: LinearUIOutputInvocation = {
-    id: LINEAR_UI_OUTPUT,
-    type: 'linear_ui_output',
-    use_cache: false,
-    is_intermediate: false,
-    board: autoAddBoardId === 'none' ? undefined : { board_id: autoAddBoardId },
+    is_intermediate: getIsIntermediate(state),
+    board: getBoardField(state),
   };
 
   const graph: NonNullableGraph = {
     id: `adhoc-esrgan-graph`,
     nodes: {
       [ESRGAN]: realesrganNode,
-      [LINEAR_UI_OUTPUT]: linearUIOutputNode,
     },
-    edges: [
-      {
-        source: {
-          node_id: ESRGAN,
-          field: 'image',
-        },
-        destination: {
-          node_id: LINEAR_UI_OUTPUT,
-          field: 'image',
-        },
-      },
-    ],
+    edges: [],
   };
 
   addCoreMetadataNode(graph, {}, ESRGAN);
